+++ CHANGELOG.md
@@ -6,12 +6,22 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+### Changed
+### Fixed
+
+## [1.17.0]
 
 ### Added
+- Added submit map button [#1179](https://github.com/PublicMapping/districtbuilder/pull/1179) & [#1201](https://github.com/PublicMapping/districtbuilder/pull/1201)
+- Added submitted column to the Organization admin projects table [#1192](https://github.com/PublicMapping/districtbuilder/pull/1192)
+- Added submission date and plan score url to organization map CSV export [#1193](https://github.com/PublicMapping/districtbuilder/pull/1193)
+- Added a lambda function to send DB server alarms to Slack [#1186](https://github.com/PublicMapping/districtbuilder/pull/1186)
 
 ### Changed
-
-### Fixed
+- Improved caching setup more & lowered cache size [#1195](https://github.com/PublicMapping/districtbuilder/pull/1195)
+- Re-enable transactions for migrations [#1200](https://github.com/PublicMapping/districtbuilder/pull/1200)
+- Drop demographics and voting fields from TopoJSON [#1197](https://github.com/PublicMapping/districtbuilder/pull/1197)
 
 ## [1.16.1]
 
@@ -480,7 +490,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release.
 
-[unreleased]: https://github.com/publicmapping/districtbuilder/compare/1.16.1...HEAD
+[unreleased]: https://github.com/publicmapping/districtbuilder/compare/1.17.0...HEAD
+[1.17.0]: https://github.com/publicmapping/districtbuilder/compare/1.16.1...1.17.0
 [1.16.1]: https://github.com/publicmapping/districtbuilder/compare/1.16.0...1.16.1
 [1.16.0]: https://github.com/publicmapping/districtbuilder/compare/1.15.1...1.16.0
 [1.15.1]: https://github.com/publicmapping/districtbuilder/compare/1.15.0...1.15.1